feat: add config for linter,  prettier, typescript, gitignore
