+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["./"],
+  "ext": "ts,json",
+  "exec": "ts-node ./index.ts",
+  "ignore": ["node_modules"]
+}
