+++ package.json
@@ -2,18 +2,26 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
+    "main": "index.ts",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start": "tsc && node ./dist/index.js",
+        "start:dev": "nodemon",
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "lint": "eslint . --ext .ts",
+        "format": "prettier --write \"**/*.ts\""
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/node": "^18.16.19",
+        "@typescript-eslint/eslint-plugin": "^5.48.2",
+        "@typescript-eslint/parser": "^5.48.2",
+        "eslint": "^8.44.0",
+        "nodemon": "^2.0.16",
+        "prettier": "^2.8.3",
+        "ts-node": "^10.9.1",
+        "typescript": "^5.0.4"
     },
     "dependencies": {
         "ws": "^8.8.0"
